updated GUI; updated writer code to reduce file size by removing trailing zeros

+++ src/cell_potts_model/CMWriter.java
@@ -6,7 +6,7 @@ public class CMWriter extends DataWriter {
 	public void writeData(CellPottsModel model, int time) {
 		writer.println();
 		int q = model.getTypesOfSpin();
-		writer.printf("%.8f ", (double) time);
+		writer.printf("%d ", time);
 		for (int i = 0; i < q; i++){
 			writer.printf("%.8f %.8f ", model.getXCM(i), model.getYCM(i));
 		}
