+++ src/cell_potts_model/EnergyWriter.java
@@ -5,7 +5,7 @@ public class EnergyWriter extends DataWriter {
 	@Override
 	public void writeData(CellPottsModel model, int time) {
 		writer.println();		
-		writer.printf("%.8f %.8f", (double) time, model.getTotalEnergy());
+		writer.printf("%d %.8f", time, model.getTotalEnergy());
 	}
 
 }
