+++ src/cell_potts_model/R2Writer.java
@@ -12,6 +12,6 @@ public class R2Writer extends DataWriter {
 	public void writeData(CellPottsModel model, int time) {
 		writer.println();	
 		double [] r2Data = model.calculateR2();
-		writer.printf("%.8f %.8f %.8f", (double) time, r2Data[0], r2Data[1]);		
+		writer.printf("%d %.8f %.8f", time, r2Data[0], r2Data[1]);		
 	}
 }
